Don't unset cloexec on server end of wayland sockets
This can prevent cosmic-panel applets closing. Presumably they didn't
receive an EOF from the socket since other applets got a copy of the
server-end fd.

As far as I can tell there's no reason to do this, for the server end of
the socket. It seems to work find with this code removed.

Fixes https://github.com/pop-os/cosmic-panel/issues/2.